move table to farms.vue
